refactor(graph): split analysis orchestration modules
